AMBARI-22899. Upgrade Infra Solr version to 7.2.1
