+++ ambari-infra/ambari-infra-solr-client/src/main/java/org/apache/ambari/infra/solr/AmbariSolrCloudClientBuilder.java
@@ -20,8 +20,7 @@
 package org.apache.ambari.infra.solr;
 
 import org.apache.solr.client.solrj.impl.CloudSolrClient;
-import org.apache.solr.client.solrj.impl.HttpClientUtil;
-import org.apache.solr.client.solrj.impl.Krb5HttpClientConfigurer;
+import org.apache.solr.client.solrj.impl.Krb5HttpClientBuilder;
 import org.apache.solr.common.cloud.SolrZkClient;
 
 public class AmbariSolrCloudClientBuilder {
@@ -31,7 +30,9 @@ public class AmbariSolrCloudClientBuilder {
   private static final String TRUSTSTORE_LOCATION_ARG = "javax.net.ssl.trustStore";
   private static final String TRUSTSTORE_PASSWORD_ARG = "javax.net.ssl.trustStorePassword";
   private static final String TRUSTSTORE_TYPE_ARG = "javax.net.ssl.trustStoreType";
-  
+  private static final String JAVA_SECURITY_AUTH_LOGIN_CONFIG = "java.security.auth.login.config";
+  private static final String SOLR_HTTPCLIENT_BUILDER_FACTORY = "solr.httpclient.builder.factory";
+
   String zkConnectString;
   String collection;
   String configSet;
@@ -45,7 +46,7 @@ public class AmbariSolrCloudClientBuilder {
   String routerField = "_router_field_";
   CloudSolrClient solrCloudClient;
   SolrZkClient solrZkClient;
-  boolean splitting;
+  boolean implicitRouting;
   String jaasFile;
   String znode;
   String saslUsers;
@@ -113,8 +114,8 @@ public class AmbariSolrCloudClientBuilder {
     return this;
   }
 
-  public AmbariSolrCloudClientBuilder withSplitting(boolean splitting) {
-    this.splitting = splitting;
+  public AmbariSolrCloudClientBuilder isImplicitRouting(boolean implicitRouting) {
+    this.implicitRouting = implicitRouting;
     return this;
   }
 
@@ -208,8 +209,8 @@ public class AmbariSolrCloudClientBuilder {
 
   private void setupSecurity(String jaasFile) {
     if (jaasFile != null) {
-      System.setProperty("java.security.auth.login.config", jaasFile);
-      HttpClientUtil.addConfigurer(new Krb5HttpClientConfigurer());
+      System.setProperty(JAVA_SECURITY_AUTH_LOGIN_CONFIG, jaasFile);
+      System.setProperty(SOLR_HTTPCLIENT_BUILDER_FACTORY, Krb5HttpClientBuilder.class.getCanonicalName());
     }
   }
 }
