+++ ambari-infra/ambari-infra-solr-plugin/src/test/java/org/apache/solr/security/InfraUserRolesLookupStrategyTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.ambari.infra.security;
+package org.apache.solr.security;
 
 import com.google.common.collect.Sets;
 import org.apache.hadoop.security.authentication.server.AuthenticationToken;
