+++ ambari-logsearch/docker/docker-compose.yml
@@ -26,7 +26,7 @@ services:
       ZOO_MY_ID: 1
       ZOO_SERVERS: server.1=zookeeper:2888:3888
   solr:
-    image: solr:${SOLR_VERSION:-6.6.2}
+    image: solr:${SOLR_VERSION:-7.2.1}
     restart: always
     hostname: solr
     ports:
